Prettier and EsLint
Ive added prettier and eslint to main so the rest of us can view and use the formatted code. ESlint should provide some rules to our code that we can follow. We can also change any of these rules if we do not like them. Also using prettier is good practice and makes the code neat so Ive added that as well. I only added one protion about the quotes but we can add more.