Prettier and ESlint
Ive added eslint and prettier to main so the rest of us can use it in our individual branches. These set rules for our code to follow and ensures our code will be uniform and neat. We can discuss any changes to these rules to ensure we all are using a style of coding that we like.